Refactored method into parent BindingVerifier
