+++ test/common/src/main/kotlin/org/codice/compliance/verification/binding/PostBindingVerifier.kt
@@ -13,7 +13,6 @@
  */
 package org.codice.compliance.verification.binding
 
-import com.google.api.client.http.HttpStatusCodes
 import de.jupf.staticlog.Log
 import io.kotlintest.matchers.shouldNotBe
 import org.apache.cxf.rs.security.saml.sso.SSOConstants.RELAY_STATE
@@ -27,7 +26,6 @@ import org.codice.compliance.SAMLBindings_3_5_4_c
 import org.codice.compliance.SAMLBindings_3_5_4_d1
 import org.codice.compliance.SAMLBindings_3_5_4_d2
 import org.codice.compliance.SAMLBindings_3_5_5_2_a
-import org.codice.compliance.SAMLBindings_3_5_6_a
 import org.codice.compliance.SAMLComplianceException
 import org.codice.compliance.SAMLProfiles_4_1_4_5
 import org.codice.compliance.allChildren
@@ -41,24 +39,7 @@ import org.codice.compliance.utils.decorators.IdpPostResponseDecorator
 import org.codice.security.saml.SamlProtocol.Binding.HTTP_POST
 import org.codice.security.sign.Decoder
 
-class PostBindingVerifier(private val response: IdpPostResponseDecorator) {
-    companion object {
-        /**
-         * Verifies the http status code of the response according to the post binding rules in the post spec
-         * 3.5.6 Error Reporting
-         */
-        fun verifyHttpStatusCode(code: Int) {
-            if (code != HttpStatusCodes.STATUS_CODE_OK) {
-                throw SAMLComplianceException.createWithPropertyMessage(
-                        SAMLBindings_3_5_6_a,
-                        property = "HTTP Status Code",
-                        actual = code.toString(),
-                        expected = "${HttpStatusCodes.STATUS_CODE_OK}"
-                )
-            }
-        }
-    }
-
+class PostBindingVerifier(private val response: IdpPostResponseDecorator) : BindingVerifier() {
     /**
      * Verify the response for a post binding
      */
